Split drag routing tests for file budget
