+++ cmuxTests/BrowserPanelTests.swift
@@ -1972,147 +1972,6 @@ final class BrowserPanelHostContainerViewTests: XCTestCase {
 
 
 @MainActor
-final class BrowserPaneDropRoutingTests: XCTestCase {
-    func testVerticalZonesFollowAppKitCoordinates() {
-        let size = CGSize(width: 240, height: 180)
-
-        XCTAssertEqual(
-            BrowserPaneDropRouting.zone(for: CGPoint(x: size.width * 0.5, y: size.height - 8), in: size),
-            .top
-        )
-        XCTAssertEqual(
-            BrowserPaneDropRouting.zone(for: CGPoint(x: size.width * 0.5, y: 8), in: size),
-            .bottom
-        )
-    }
-
-    func testTopChromeHeightPushesTopSplitThresholdIntoWebView() {
-        let size = CGSize(width: 240, height: 180)
-
-        XCTAssertEqual(
-            BrowserPaneDropRouting.zone(
-                for: CGPoint(x: size.width * 0.5, y: 110),
-                in: size,
-                topChromeHeight: 36
-            ),
-            .center
-        )
-        XCTAssertEqual(
-            BrowserPaneDropRouting.zone(
-                for: CGPoint(x: size.width * 0.5, y: 150),
-                in: size,
-                topChromeHeight: 36
-            ),
-            .top
-        )
-    }
-
-    func testHitTestingCapturesOnlyForRelevantDragEvents() {
-        XCTAssertTrue(
-            BrowserPaneDropTargetView.shouldCaptureHitTesting(
-                pasteboardTypes: [DragOverlayRoutingPolicy.bonsplitTabTransferType],
-                eventType: .cursorUpdate
-            )
-        )
-        XCTAssertFalse(
-            BrowserPaneDropTargetView.shouldCaptureHitTesting(
-                pasteboardTypes: [DragOverlayRoutingPolicy.bonsplitTabTransferType],
-                eventType: .leftMouseDown
-            )
-        )
-        XCTAssertFalse(
-            BrowserPaneDropTargetView.shouldCaptureHitTesting(
-                pasteboardTypes: [.fileURL],
-                eventType: .cursorUpdate
-            )
-        )
-
-        let externalPayloads: [[NSPasteboard.PasteboardType]] = [
-            [.fileURL],
-            [.URL],
-            [.png],
-            [.tiff],
-            [.html],
-            [.string],
-            [.fileURL, .png],
-        ]
-
-        for pasteboardTypes in externalPayloads {
-            XCTAssertFalse(
-                BrowserPaneDropTargetView.shouldCaptureHitTesting(
-                    pasteboardTypes: pasteboardTypes,
-                    eventType: .cursorUpdate
-                ),
-                "Browser pane drop target should not capture external drag payload: \(pasteboardTypes)"
-            )
-        }
-    }
-
-    func testCenterDropOnSamePaneIsNoOp() {
-        let paneId = PaneID(id: UUID())
-        let target = BrowserPaneDropContext(
-            workspaceId: UUID(),
-            panelId: UUID(),
-            paneId: paneId
-        )
-        let transfer = BrowserPaneDragTransfer(
-            tabId: UUID(),
-            sourcePaneId: paneId.id,
-            sourceProcessId: Int32(ProcessInfo.processInfo.processIdentifier)
-        )
-
-        XCTAssertEqual(
-            BrowserPaneDropRouting.action(for: transfer, target: target, zone: .center),
-            .noOp
-        )
-    }
-
-    func testRightEdgeDropBuildsSplitMoveAction() {
-        let paneId = PaneID(id: UUID())
-        let target = BrowserPaneDropContext(
-            workspaceId: UUID(),
-            panelId: UUID(),
-            paneId: paneId
-        )
-        let tabId = UUID()
-        let transfer = BrowserPaneDragTransfer(
-            tabId: tabId,
-            sourcePaneId: UUID(),
-            sourceProcessId: Int32(ProcessInfo.processInfo.processIdentifier)
-        )
-
-        XCTAssertEqual(
-            BrowserPaneDropRouting.action(for: transfer, target: target, zone: .right),
-            .move(
-                tabId: tabId,
-                targetWorkspaceId: target.workspaceId,
-                targetPane: paneId,
-                splitTarget: BrowserPaneSplitTarget(orientation: .horizontal, insertFirst: false)
-            )
-        )
-    }
-
-    func testDecodeTransferPayloadReadsTabAndSourcePane() {
-        let tabId = UUID()
-        let sourcePaneId = UUID()
-        let payload = try! JSONSerialization.data(
-            withJSONObject: [
-                "tab": ["id": tabId.uuidString],
-                "sourcePaneId": sourcePaneId.uuidString,
-                "sourceProcessId": ProcessInfo.processInfo.processIdentifier,
-            ]
-        )
-
-        let transfer = BrowserPaneDragTransfer.decode(from: payload)
-
-        XCTAssertEqual(transfer?.tabId, tabId)
-        XCTAssertEqual(transfer?.sourcePaneId, sourcePaneId)
-        XCTAssertTrue(transfer?.isFromCurrentProcess == true)
-    }
-}
-
-
-@MainActor
 final class WindowBrowserSlotViewTests: XCTestCase {
     private final class CapturingView: NSView {
         override func hitTest(_ point: NSPoint) -> NSView? {
